+++ CHANGELOG.md
@@ -29,6 +29,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.0.0] - 2023-01-24
+
+### Added
+
+- Introduce `AesEncryption()` to support other AES encryption modes:
+    - `AES/CBC/PKCS5Padding`
+    - `AES/CTR/NoPadding`
+    - `"AES/CFB/NoPadding`
+    - `AES/OFB/NoPadding`
+
+### Changed
+
+- `AesGcmEncryption()` was renamed to `AesEncryption()`
+
+### Removed
+
+- BREAKING CHANGE: `AesGcmEncryption()` is no longer available. Use `AesEncryption()` instead.
+
 ## [1.1.0] - 2023-01-11
 
 ### Added